+++ resources/android/src/com/victorycodedev/plugins/toastkit/model/ToastKitModels.kt
@@ -31,6 +31,9 @@ internal data class ToastKitStyle(
 
 internal data class ToastKitConfiguration(
     val id: String,
+    val uniqueKey: String?,
+    val nativeIos: Boolean,
+    val nativeAndroid: Boolean,
     val message: String,
     val title: String?,
     val variant: String,