+++ resources/ios/Sources/Model/ToastKitModels.swift
@@ -38,6 +38,9 @@ struct ToastKitStyleConfiguration {
 
 struct ToastKitConfiguration: Identifiable {
     let id: String
+    let uniqueKey: String?
+    var nativeIos: Bool
+    var nativeAndroid: Bool
     var message: String
     var title: String?
     var variant: String