+++ resources/js/index.js
@@ -1,6 +1,20 @@
 export { Toast } from "./Toast.js";
 export { PendingToast } from "./PendingToast.js";
 export { PendingToastUpdate } from "./PendingToastUpdate.js";
-export { Show, Update, Dismiss, DismissAll } from "./toastkit.js";
-export { show, update, dismiss, dismissAll } from "./toastkit.js";
+export {
+  Show,
+  Update,
+  UpdateUnique,
+  Dismiss,
+  DismissUnique,
+  DismissAll,
+} from "./toastkit.js";
+export {
+  show,
+  update,
+  updateUnique,
+  dismiss,
+  dismissUnique,
+  dismissAll,
+} from "./toastkit.js";
 export { Toast as default } from "./Toast.js";